bot/requirements.txt
aiogram>=3.4.0
python-dotenv>=1.0.0
openai>=1.40.0
anthropic>=0.36.0
httpx>=0.27.0
uvicorn>=0.30.0

